+++ packages/nodes-base/nodes/Evaluation/EvaluationTrigger/EvaluationTrigger.node.json
